requirements.txt
modelscope
torch
torchvision
Pillow
flask
requests
diffusers
transformers
accelerate
gunicorn
gevent
